Possibly didn't not undo a broken fix for a working problem that was simply confusing with conditions. I think it has something to do with the EoL format, but basically, conditions weren't updating properly.
I really don't know what I did. The changes around line 3756 were after it was working just to improve the aesthetics of the generated code. The change around  line 4219 was to undo a previous "fix" to what I think is the root cause of the issue. I just don't know what that cause is. It works now with the old code, though, so I think everything works.